fixing warnings, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@18552 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netbuild/NBEdge.cpp
@@ -1650,7 +1650,7 @@ NBEdge::divideOnEdges(const EdgeVector* outgoing, const bool buildCrossingsAndWa
     for (int i = 0; i < numOutgoing; i++) {
         // tmpNo will be the number of connections from this edge
         //  to the next edge
-        assert(i < resultingLanes.size());
+        assert(i < (int)resultingLanes.size());
         const SUMOReal tmpNum = resultingLanes[i] / minResulting;
         for (SUMOReal j = 0; j < tmpNum; j++) {
             transition.push_back((*outgoing)[i]);
